Create an internaltypes.h header file.  Add printStack and printEquivalenceMap to debug.cpp.
