+++ src/debug.cpp
@@ -18,4 +18,32 @@ limitations under the License.
 
 namespace libcellml {
 
+void printStack(const IndexStack &stack)
+{
+    Debug(false) << "[";
+    for (auto iter = stack.begin(); iter < stack.end(); ++iter) {
+        Debug(false) << *iter;
+        if (iter + 1 < stack.end()) {
+            Debug(false) << ", ";
+        }
+    }
+    Debug() << "]";
+}
+
+void printEquivalenceMap(const EquivalenceMap &map)
+{
+    Debug() << "Print out of equivalence map";
+    for (EquivalenceMap::const_iterator iter = map.begin(); iter != map.end(); ++iter) {
+        auto key = iter->first;
+        Debug(false) << "key: ";
+        printStack(key);
+        auto vector = iter->second;
+        for (auto vectorIt = vector.begin(); vectorIt < vector.end(); ++vectorIt) {
+            Debug(false) << "value: ";
+            printStack(*vectorIt);
+        }
+    }
+}
+
+
 }
