+++ src/debug.h
@@ -19,6 +19,8 @@ limitations under the License.
 #include <iostream>
 #include <sstream>
 
+#include "internaltypes.h"
+
 namespace libcellml {
 
 struct Debug
@@ -56,5 +58,8 @@ private:
     std::ostringstream mSS;
     bool mNewLine;
 };
+
+void printStack(const IndexStack &stack);
+void printEquivalenceMap(const EquivalenceMap &map);
 
 } // namespace libcellml
